+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+
+
+## [0.0.3] - 2026-08-10
+
+### Added
+
+- implemented proper SSE streaming and event formatting
+- user messages now can hold file attachments (flask db upgrade)
+- implemented write file tool (needs testing)
+- implemented send mail tool
+
 ## [0.0.2] - 2026-08-06
 
 ### Added
@@ -31,6 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - error handling and debug utils
 - llm service layer vor gemini
 
-[Unreleased]: https://github.com/negsi/trinity-flask/compare/v0.0.2...develop
+[Unreleased]: https://github.com/negsi/trinity-flask/compare/v0.0.3...develop
+[0.0.3]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.3
 [0.0.2]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.2
 [0.0.1]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.1